+++ srcs/structs/signals.c
